tests/test_module_06_api.py
"""Module 6: FastAPI Deployment - Tests for REST API."""

import pytest
from fastapi.testclient import TestClient


class TestFastAPIDeployment:
    """Test suite for Module 6: FastAPI Deployment"""
    
    def test_api_imports(self):
        """TODO 6.0: Verify API module can be imported."""
        try:
            from src.api.main import app
        except ImportError as e:
            pytest.fail(f"Failed to import FastAPI app: {e}")
    
    def test_app_instance_created(self):
        """TODO 6.1: Test FastAPI app instance is created."""
        pytest.skip("TODO: Implement test after completing TODO 6.1")
        
        from src.api.main import app
        from fastapi import FastAPI
        
        assert isinstance(app, FastAPI)
        assert app.title is not None
    
    def test_health_endpoint(self):
        """TODO 6.4: Test /health endpoint returns 200."""
        pytest.skip("TODO: Implement test after completing TODO 6.4")
        
        from src.api.main import app
        
        client = TestClient(app)
        response = client.get("/health")
        
        assert response.status_code == 200
        assert response.json()["status"] == "healthy"
    
    def test_predict_endpoint_exists(self):
        """TODO 6.5: Test /predict endpoint exists and accepts files."""
        pytest.skip("TODO: Implement test after completing TODO 6.5")
        
        from src.api.main import app
        import io
        from PIL import Image
        
        client = TestClient(app)
        
        # Create dummy image
        image = Image.new('RGB', (256, 256), color='red')
        image_bytes = io.BytesIO()
        image.save(image_bytes, format='PNG')
        image_bytes.seek(0)
        
        response = client.post(
            "/predict",
            files={"file": ("test.png", image_bytes, "image/png")}
        )
        
        assert response.status_code == 200
    
    def test_predict_response_format(self):
        """TODO 6.5: Test /predict response matches Pydantic model."""
        pytest.skip("TODO: Implement test after completing TODO 6.5")
        
        from src.api.main import app, PredictionResponse
        import io
        from PIL import Image
        
        client = TestClient(app)
        
        image = Image.new('RGB', (256, 256), color='blue')
        image_bytes = io.BytesIO()
        image.save(image_bytes, format='PNG')
        image_bytes.seek(0)
        
        response = client.post(
            "/predict",
            files={"file": ("test.png", image_bytes, "image/png")}
        )
        
        data = response.json()
        
        # Should have expected fields
        assert "prediction" in data or "mask" in data
        assert "confidence" in data or "inference_time" in data
    
    def test_invalid_file_format(self):
        """TODO 6.6: Test API handles invalid file formats."""
        pytest.skip("TODO: Implement test after completing TODO 6.6")
        
        from src.api.main import app
        
        client = TestClient(app)
        
        # Send text file instead of image
        response = client.post(
            "/predict",
            files={"file": ("test.txt", b"not an image", "text/plain")}
        )
        
        assert response.status_code == 400
    
    def test_model_lazy_loading(self):
        """TODO 6.3: Test model is loaded lazily (on first request)."""
        pytest.skip("TODO: Implement test after completing TODO 6.3")
        
        from src.api.main import app, get_model
        
        # Model should not be loaded at startup
        # This is hard to test without modifying code
        # Just verify the function exists
        assert callable(get_model)
    
    def test_pydantic_request_validation(self):
        """TODO 6.2: Test Pydantic models validate requests."""
        pytest.skip("TODO: Implement test after completing TODO 6.2")
        
        from src.api.main import PredictionRequest
        from pydantic import ValidationError
        
        # Valid request
        valid = PredictionRequest(image_base64="dGVzdA==", threshold=0.5)
        assert valid.threshold == 0.5
        
        # Invalid request (wrong type)
        with pytest.raises(ValidationError):
            PredictionRequest(image_base64="dGVzdA==", threshold="invalid")
    
    def test_error_handling(self):
        """TODO 6.6: Test API returns proper error codes."""
        pytest.skip("TODO: Implement test after completing TODO 6.6")
        
        from src.api.main import app
        
        client = TestClient(app)
        
        # Request to non-existent endpoint
        response = client.get("/nonexistent")
        assert response.status_code == 404
        
        # Invalid JSON (if applicable)
        response = client.post("/predict", data="invalid json")
        assert response.status_code in [400, 422]
