requirements.txt
# requirements.txt - Python dependencies for Remote Sensing ML Tutorial

# Core deep learning
torch>=2.0.0
torchvision>=0.15.0

# Geospatial data processing
rasterio>=1.3.0
xarray>=2023.1.0
numpy>=1.24.0
pillow>=10.0.0

# API framework
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Monitoring
prometheus-client>=0.17.0

# Model export and optimization
onnx>=1.14.0
onnxruntime>=1.15.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.24.0  # For FastAPI TestClient

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Optional: For enhanced development experience
black>=23.0.0  # Code formatting
isort>=5.12.0  # Import sorting
mypy>=1.5.0    # Type checking
